created Post component instead of seperated components
