+++ src/pages/Home.jsx
@@ -1,16 +1,10 @@
 import {
-  TextPost,
   AwardPost,
-  BildTextPost,
-  ShortTextPost,
-  VideoPost,
-  AudioPost,
   UmfragePost
 } from "../posts/Posttypes";
+import {Post} from "../components/Post"
 import { useContext } from "react";
 import { PostContext } from "../App";
-import Navbar from "../components/Navbar";
-import { Navigate } from "react-router-dom";
 const Home = () => {
   const { postData, database } = useContext(PostContext);
 
@@ -22,17 +16,13 @@ const Home = () => {
         database.map((post,index)=>{
           switch (post.postType) {
             case "picture":
-              return <BildTextPost post={post} index={index}/>
             case "text":
-              return <TextPost post={post} index={index}/>
             case "shortText":
-              return <ShortTextPost post={post} index={index}/>
+            case "video":
+            case "audio":
+              return <Post post={post} index={index}/>
             case "awardPost":
               return <AwardPost post={post} index={index}/>
-            case "video":
-              return <VideoPost post={post} index={index}/>
-            case "audio":
-              return <AudioPost post={post} index={index}/>
             case "survey":
               return <UmfragePost post={post} index={index}/>
             default:
